+++ testFiles/useNsModule.js
@@ -1,0 +1,9 @@
+define(['uiComponent'], function (Component) {
+  return Component.extend({
+    initialize: function () {
+      this.useNs(function (ns) {
+        return ns.statusOptions
+      })
+    }
+  })
+})